requirements.txt
CairoSVG==2.4.2
matplotlib==3.3.2
networkx==2.5
numpy==1.19.4
opt-einsum==3.3.0
orderedset==2.0.3
paddlepaddle-gpu==2.6.1.post112
pandas==1.1.3
rdkit==2022.3.3
scipy==1.5.0
seaborn==0.11.0
tabulate==0.8.7
tqdm==4.50.0
